auth_service/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
psycopg2-binary
alembic
python-jose[cryptography]
httpx
pydantic
python-dotenv
pydantic-settings
